Implemented stub :queue indirection terminus for catalog class, with accompanying bare-bones tests.
This introduces Puppet::Node::Catalog::Queue for indirection of catalog storage to a message queue
via the abstract :queue terminus type.

This stub is presently non-functional, because the to_message and from_message methods (which it
overrides) don't do anything.  They ultimately should be adjusted to use Marshal for (de)serialization
of the catalog object(s) for traversing the message queue.  The unit test should be extended to
verify the use of Marshal; YAML may have problems at present, and certainly is likely to be fairly slow
compared to Marshal.