Initial implementation of composite "Store Configs" class.
This class may end up being throw-away, since the current design of storeconfigs
muddies the design waters with nodes, facts, and catalog.  For the purpose of getting
indirection working quickly with store configs, however, this class gives us a way
forward.

Puppet::Node::StoreConfig represents the data you need for a store config
operation: the node, and the catalog.  Under the hood, it only preserves the
catalog's vertices and ignores the rest of the catalog object provided.

The accompanying indirection classes allow for indirection with :queue and
:active_record.  The :active_record indirection's behavior is more or less consistent
with the original implementation: it calls out to Puppet::Rails::Node.store() to
save to the database.