Fixing #2149 - Facts are passed as part of the catalog request
This removes the requirement of shared fact caching
on the servers, since the server responding to the catalog
request will receive the facts as part of the request.

The facts are serialized as a parameter to the request,
rather than each being set as a separate request parameter.

This hard-codes yaml as the serialization format for the
facts, because I couldn't get marshal to work and it's just not
as big a deal for such a small amount of data.

Signed-off-by: Luke Kanies <luke@madstop.com>