Fixing #2139 - Modules now search across all module paths
If you have a multi-directory module path, then a given
module with now search through all of the directories of
that name in all of your module paths.

Thus, for instance, you can have 'dist' and 'site' module
directories, and you can override a configuration file for a
module in your 'dist' directory by putting a new file
in your 'site' directory.

This is done to enable #2140 - my plan is to add support
for external data files, and you would be able to override
the data files by dropping a supplementary data file in
your 'site' module directory.

Signed-off-by: Luke Kanies <luke@madstop.com>